[EDIFIKANA] Improving the Staff List screen
