+++ edifikana/front-end/shared-app/src/commonMain/kotlin/com/cramsan/edifikana/client/lib/features/management/hub/HubScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.Modifier
 import com.cramsan.edifikana.client.lib.features.management.drawer.ManagementViewModel
 import com.cramsan.edifikana.client.lib.features.management.home.AccountDropDown
 import com.cramsan.edifikana.client.lib.features.management.properties.PropertyManagerScreen
-import com.cramsan.edifikana.client.lib.features.management.timecardstafflist.StaffListScreen
+import com.cramsan.edifikana.client.lib.features.management.stafflist.StaffListScreen
 import com.cramsan.edifikana.client.ui.components.EdifikanaTopBar
 import edifikana_lib.Res
 import edifikana_lib.app_name
