+++ edifikana/front-end/shared-app/src/commonMain/kotlin/com/cramsan/edifikana/client/lib/features/management/staff/StaffScreenPreview.kt
@@ -11,15 +11,37 @@ import org.jetbrains.compose.ui.tooling.preview.Preview
  */
 @Preview
 @Composable
-private fun StaffScreenPreview() = AppTheme {
+private fun StaffScreenPreview_Primary_Staff() = AppTheme {
+    StaffContent(
+        content = StaffUIState(
+            title = "Cesar Vargas",
+            isLoading = true,
+            idType = null,
+            idNNumber = null,
+            firstName = "Cesar",
+            lastName = "Vargas",
+            email = "admin@cenit.com",
+            role = null,
+            isEditable = false,
+        ),
+        onBackSelected = {},
+    )
+}
+
+@Preview
+@Composable
+private fun StaffScreenPreview_Secondary_Staff() = AppTheme {
     StaffContent(
         content = StaffUIState(
             title = "Cesar Vargas",
             isLoading = true,
             idType = IdType.DNI,
+            idNNumber = "12345678",
             firstName = "Cesar",
             lastName = "Vargas",
+            email = "admin@cenit.com",
             role = StaffRole.SECURITY_COVER,
+            isEditable = true,
         ),
         onBackSelected = {},
     )
