+++ edifikana/front-end/shared-app/src/commonMain/kotlin/com/cramsan/edifikana/client/lib/features/management/stafflist/StaffListEvent.kt
@@ -3,7 +3,10 @@ package com.cramsan.edifikana.client.lib.features.management.stafflist
 import com.cramsan.framework.core.compose.ViewModelEvent
 
 /**
- * Represents the UI state of the Staff List screen.
+ * Events that can be triggered within the domain of the StaffList feature.
+ *
+ * Events are triggered from a ViewModel and are consumed by the UI.
+ *
  */
 sealed class StaffListEvent : ViewModelEvent {
     /**
