+++ edifikana/front-end/shared-app/src/commonMain/kotlin/com/cramsan/edifikana/client/lib/features/management/stafflist/StaffListUIState.kt
@@ -1,20 +1,32 @@
 package com.cramsan.edifikana.client.lib.features.management.stafflist
 
+import com.cramsan.edifikana.lib.model.StaffId
+import com.cramsan.edifikana.lib.model.StaffStatus
 import com.cramsan.framework.core.compose.ViewModelUIState
 
 /**
- * Represents the UI state of the Staff List screen.
+ * UI state of the StaffList feature.
+ *
+ * This class models the top level state of the page.
+ * For modeling more specific details of the page, see the respective UI model class.
  */
 data class StaffListUIState(
-    val staffs: List<com.cramsan.edifikana.client.lib.features.management.stafflist.StaffUIModel>,
     val isLoading: Boolean,
-    val title: String,
+    val staffList: List<StaffUIModel>,
 ) : ViewModelUIState {
     companion object {
-        val Empty = StaffListUIState(
-            emptyList(),
-            true,
-            "",
-        )
+        val Initial = StaffListUIState(true, emptyList())
     }
 }
+
+/**
+ * UI model for a staff member.
+ *
+ * This class models the data that is displayed in the staff list.
+ */
+data class StaffUIModel(
+    val id: StaffId,
+    val name: String,
+    val email: String?,
+    val status: StaffStatus,
+)
