+++ edifikana/front-end/shared-app/src/commonMain/kotlin/com/cramsan/edifikana/client/lib/features/management/stafflist/StaffListViewModel.kt
@@ -3,74 +3,84 @@ package com.cramsan.edifikana.client.lib.features.management.stafflist
 import com.cramsan.edifikana.client.lib.features.management.ManagementDestination
 import com.cramsan.edifikana.client.lib.features.window.EdifikanaWindowsEvent
 import com.cramsan.edifikana.client.lib.managers.StaffManager
+import com.cramsan.edifikana.client.lib.models.StaffModel
 import com.cramsan.edifikana.lib.model.StaffId
 import com.cramsan.framework.core.compose.BaseViewModel
 import com.cramsan.framework.core.compose.ViewModelDependencies
-import com.cramsan.framework.core.compose.resources.StringProvider
-import edifikana_lib.Res
-import edifikana_lib.title_timecard_staff_list
 import kotlinx.coroutines.launch
 
 /**
- * Represents the UI state of the Staff List screen.
- */
+ * ViewModel for the StaffList screen.
+ **/
 class StaffListViewModel(
-    private val staffManager: StaffManager,
-    private val stringProvider: StringProvider,
     dependencies: ViewModelDependencies,
+    private val staffManager: StaffManager,
 ) : BaseViewModel<StaffListEvent, StaffListUIState>(
     dependencies,
-    StaffListUIState.Empty,
-    TAG
+    StaffListUIState.Initial,
+    TAG,
 ) {
 
     /**
-     * Load staff members.
+     * Load the staff list.
      */
-    fun loadStaffs() = viewModelScope.launch {
-        updateUiState { it.copy(isLoading = true) }
-        val result = staffManager.getStaffList()
+    fun loadStaffList() {
+        viewModelScope.launch {
+            val staffList = staffManager.getStaffList().onFailure {
+                updateUiState {
+                    it.copy(isLoading = false)
+                }
+                return@launch
+            }
+                .getOrThrow()
+                .map { it.toUIModel() }
 
-        if (result.isFailure) {
-            val state =
-                StaffListUIState(
-                    emptyList(),
-                    false,
-                    stringProvider.getString(Res.string.title_timecard_staff_list),
-                )
-            updateUiState { state }
-        } else {
-            val staffs = result.getOrThrow()
-            val state =
-                StaffListUIState(
-                    staffs.map {
-                        it.toUIModel()
-                    },
-                    false,
-                    stringProvider.getString(Res.string.title_timecard_staff_list)
-                )
-            updateUiState { state }
+            updateUiState {
+                it.copy(isLoading = false, staffList = staffList)
+            }
+        }
+    }
+
+    private fun StaffModel.toUIModel(): StaffUIModel {
+        return StaffUIModel(
+            id = id,
+            name = name,
+            email = email,
+            status = status,
+        )
+    }
+
+    /**
+     * Navigate to the AddPrimaryStaff screen.
+     */
+    fun navigateToAddPrimaryStaff() {
+        viewModelScope.launch {
+            emitWindowEvent(
+                EdifikanaWindowsEvent.NavigateToScreen(ManagementDestination.AddPrimaryStaffManagementDestination)
+            )
         }
     }
 
     /**
-     * Navigate to staff member.
+     * Navigate to the AddSecondaryStaff screen.
      */
-    fun navigateToStaff(staffPK: StaffId) = viewModelScope.launch {
-        emitWindowEvent(
-            EdifikanaWindowsEvent.NavigateToScreen(
-                ManagementDestination.TimeCardSingleStaffDestination(staffPK)
+    fun navigateToAddSecondaryStaff() {
+        viewModelScope.launch {
+            emitWindowEvent(
+                EdifikanaWindowsEvent.NavigateToScreen(ManagementDestination.AddSecondaryStaffManagementDestination)
             )
-        )
+        }
     }
 
     /**
-     * Navigate back.
+     * Navigate to the staff screen.
      */
-    fun navigateBack() = viewModelScope.launch {
-        emitWindowEvent(
-            EdifikanaWindowsEvent.NavigateBack
-        )
+    fun navigateToStaff(staffId: StaffId) {
+        viewModelScope.launch {
+            emitWindowEvent(
+                EdifikanaWindowsEvent.NavigateToScreen(ManagementDestination.StaffDestination(staffId))
+            )
+        }
     }
 
     companion object {
