+++ edifikana/front-end/shared-app/src/commonMain/kotlin/com/cramsan/edifikana/client/lib/features/management/timecardstafflist/TimeCardStaffUIModel.kt
@@ -1,13 +1,13 @@
-package com.cramsan.edifikana.client.lib.features.management.stafflist
+package com.cramsan.edifikana.client.lib.features.management.timecardstafflist
 
 import com.cramsan.edifikana.client.lib.models.StaffModel
 import com.cramsan.edifikana.client.lib.models.fullName
 import com.cramsan.edifikana.lib.model.StaffId
 
 /**
- * Represents the UI state of the Staff List screen.
+ * Represents the UI state of the Time Card Staff List screen.
  */
-data class StaffUIModel(
+data class TimeCardStaffUIModel(
     val fullName: String,
     val staffPK: StaffId,
 )
@@ -15,8 +15,8 @@ data class StaffUIModel(
 /**
  * Represents the UI state of the Staff List screen.
  */
-fun StaffModel.toUIModel(): StaffUIModel {
-    return StaffUIModel(
+fun StaffModel.toUIModel(): TimeCardStaffUIModel {
+    return TimeCardStaffUIModel(
         fullName = fullName(),
         staffPK = id,
     )
